Added animal filter

+++ Saponja.Data/Entities/Models/Geolocation.cs
@@ -9,7 +9,7 @@ namespace Saponja.Data.Entities.Models
     [Owned]
     public class Geolocation
     {
-        public float Longitude { get; set; }
-        public float Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
